+++ app/nusim.sh
@@ -9,10 +9,17 @@
 
 newsim()
 {
-    secho "# nusim: $@" 'menu'
+    local title="# nusim: $@"
+    secho "${title}" 'menu'
     startTime=$(date '+%Y-%m-%d %H:%M');
-    secho "${startTime}" 'blue'
+    secho "${startTime}" 'green'
     sleepUnit=3
+    local scope="ps"
+    if [ ! -z $2 ]; then
+        scope="core"
+    fi
+
+    newStep 'setVar'
 
     case ${1} in
         'check')
@@ -25,17 +32,17 @@ newsim()
             unixInstall git
             unixInstall composer
 
-            local gitRepoPath="${_nusim_tmppath}/repo"
+            local gitRepoPath="${tmp}/repo"
             rm -rf ${gitRepoPath}/nusim
             mkdir -p ${gitRepoPath}
             cd ${gitRepoPath}
             ssudo "rm /usr/local/bin/nusim"
             secho "# git clone ${_git_GitHub_nusim} ${gitRepoPath}" 'menu'
-            git clone -b develop --recursive ${_git_GitHub_nusim} ./nusim
+            git clone -b ${_nusim_install_branch} --recursive ${_git_GitHub_nusim} ./nusim
             cd nusim
             composer install
-            ssudo "touch ${_nusim_tmppath}/nusim.log"
-            ssudo "chmod 777 ${_nusim_tmppath}/nusim.log"
+            ssudo "touch ${tmp}/nusim.log"
+            ssudo "chmod 777 ${tmp}/nusim.log"
             secho "* install nusim" 'menu'
             ssudo "php bin/nusim self:compile -e dev -i -v"
             nusim --list
@@ -43,175 +50,161 @@ newsim()
         ;;
 
         'installInstance')
-            setYmlVal "_nusim_sugar_db_key" 'dbKey'
-            local cmd="nusim install:developer -e dev --license-key ${_nusim_sugar_license} --repo-path ${_nusim_sugar_mango}"
-            cmd="${cmd} --instance ${_nusim_sugar_name} --sugar-version ${_nusim_sugar_version} --sugar-flavor=${_nusim_sugar_flavor}"
-            cmd="${cmd} --admin-user-name ${_nusim_sugar_admin_user} --admin-password ${_nusim_sugar_admin_pass}"
+            cmdConcat "install:developer -e dev --license-key ${license} --repo-path ${mango}"
+            cmdConcat "--instance ${build}${suffix} --sugar-version ${version} --sugar-flavor=${flavor}"
+            cmdConcat "--admin-user-name ${adminName} --admin-password ${adminPass}"
+            newStep 'dbKeyUser'
 
-            if [ ${dbKey} == 'oracle' ]; then
-                cmd="${cmd} --db-type ${_db_oracle_type} --db-user ${_db_oracle_connect_user} --db-pass ${_db_oracle_connect_pass}"
-                cmd="${cmd} --db-host ${_db_oracle_connect_host} --db-port ${_db_oracle_port} --db-name ${_db_oracle_setRoot_host}/orcl"
-            else
-                cmd="${cmd} --db-type ${_db_mysql_type} --db-user ${_db_mysql_connect_user} --db-pass ${_db_mysql_connect_pass}"
-                cmd="${cmd} --db-host ${_db_mysql_connect_host} --db-port ${_db_mysql_port} --db-name turbinado"
-            fi
-            secho "${cmd}" 'menu'
-            ${cmd} &
-            ssudo 'tail -f /tmp/nusim/nusim.log'
+            tailPidCmd "${cmd}" "${tmp}/nusim.log"
+
+            wait $cmdPID
+            sleep $sleepUnit
+            secho "nusim command finished" 'menu'
+            
             local localIpCmd="ip -f inet addr show eth1 | grep -Eo 'inet (addr:)?([0-9]*\.){3}[0-9]*' | grep -Eo '([0-9]*\.){3}[0-9]*'"
             local ipVal=$(eval "${localIpCmd}")
-            secho "http://${ipVal}/${_nusim_sugar_name}/${_nusim_sugar_flavor}" 'green'
+            secho "http://${ipVal}/${sugarName}/${flavor}" 'green'
         ;;
 
         'createInstallPack')
-            local dockerImageName="build_core_install_${_nusim_build_number}"
-            if [ "${_nusim_build_with_nomad}" == 'true' ]; then
-                gitCloneOrUpdate ${_git_GitHub_nomad} ${_nusim_tmppath}/repo nomad ${_nusim_build_branch_nomad} ${_nusim_build_checkout_nomad}
-            fi
-            gitCloneOrUpdate ${_git_GitHub_translations} ${_nusim_tmppath}/repo translations ${_nusim_build_branch_translations} ${_nusim_build_checkout_translations}
-            gitCloneOrUpdate ${_git_GitHub_refinery} ${_nusim_tmppath}/repo refinery ${_nusim_build_branch_refinery} ${_nusim_build_checkout_refinery}
+            cmdConcat "package:create:${scope}:install -e dev"
+            cmdConcat "--mango-path ${mango}"
+            cmdConcat "--translations-path ${tmp}/repo/translations"
+            cmdConcat "--refinery-path ${tmp}/repo/refinery"
+            cmdConcat "--build-path ${tmp}/builds/${build}"
+            cmdConcat "--build-number ${build}"
+            cmdConcat "--sugar-version ${version}"
+            cmdConcat "--config-version ${configv}"
+            cmdConcat "--sugar-flavor ${flavor}"
 
-            cd ${_nusim_sugar_mango}
+            newStep 'gitCloneOrUpdate' 'refinery'
+            if [ "${useN}" == "true" ]; then
+                newStep 'gitCloneOrUpdate' 'nomad'
+                cmdConcat "--nomad-path=${tmp}/repo/nomad"
+            fi
+            if [ "${useT}" == "true" ]; then
+                newStep 'gitCloneOrUpdate' 'translations'
+            else
+                cmdConcat "--no-latin"
+            fi
+            if [ "${useMC}" == "true" ]; then
+                newStep 'gitCloneOrUpdate' 'mangoCore'
+            fi
+
+            cd ${mango}
             gitMango postCheckout
 
-            local cmd="nusim package:create:ps:install -e dev --mango-path ${_nusim_sugar_mango}"
-            if [ "${_nusim_build_with_nomad}" == 'true' ]; then
-                cmd="${cmd} --nomad-path=${_nusim_tmppath}/repo/nomad"
-            fi
-            cmd="${cmd} --translations-path ${_nusim_tmppath}/repo/translations"
-            cmd="${cmd} --refinery-path ${_nusim_tmppath}/repo/refinery"
-            cmd="${cmd} --build-path ${_nusim_tmppath}/builds/${_nusim_build_number}"
-            cmd="${cmd} --build-number ${_nusim_build_number}"
-            cmd="${cmd} --sugar-version ${_nusim_sugar_version} --sugar-flavor ${_nusim_sugar_flavor}"
+            nusim -V
             if [ "$ME" == 'vagrant' ]; then
                 cmd="echo \"${cmd}\" | sudo su -"
             fi
-
-            nusim -V
-            tailPidCmd "${cmd}" '/tmp/nusim/nusim.log'
+            tailPidCmd "${cmd}" "${tmp}/nusim.log"
 
             local sleepCount=0
-            while [ -z "$(docker ps -a | grep build_core_install_)" -a $sleepCount -lt 300 ]; do
+            while [ -z "$(docker ps -a | grep build_core_)" -a $sleepCount -lt 300 ]; do
                 sleep $sleepUnit
                 let sleepCount=sleepCount+sleepUnit
             done
 
-            cmd="docker logs -f ${dockerImageName}"
-            if [ "$ME" == 'vagrant' ]; then
-                secho "echo '${cmd} &' | sudo su -" 'menu'
-                echo "${cmd}" | sudo su -
-            else
-                secho "${cmd}" 'menu'
-                ${cmd}
-            fi
+            newStep 'logDocker'
 
             wait $cmdPID
             sleep $sleepUnit
             secho "nusim command finished" 'menu'
 
-            nusimZipBuild ${_nusim_build_number}
-            nusimCpFullBuild ${_nusim_build_number}
+            nusimZipBuild ${build}
+            nusimCpFullBuild ${build}
 
             secho "/** Finished Create Install Pack **/" 'menu'
         ;;
 
         'createUpgradePack')
-            local dockerImageName="build_core_install_${_nusim_build_upgrade_number}"
-            local pathToBaseBuildPack="${_nusim_tmppath}/builds/${_nusim_build_number}"
+            cmdConcat "package:create:ps:upgrade -e dev"
+            cmdConcat "--mango-path ${mango}"
+            cmdConcat "--translations-path ${tmp}/repo/translations"
+            cmdConcat "--refinery-path ${tmp}/repo/refinery"
+            cmdConcat "--baseline-path ${tmp}/builds/${build}"
+            cmdConcat "--build-path ${tmp}/builds/${upgradePack}"
+            cmdConcat "--build-number ${upgradePack}"
+            cmdConcat "--sugar-version ${version} "
+            cmdConcat "--config-version ${configv}"
+            cmdConcat "--sugar-flavor ${flavor}"
 
-            if [ "${_nusim_build_with_nomad}" == 'true' ]; then
-                gitCloneOrUpdate ${_git_GitHub_nomad} ${_nusim_tmppath}/repo nomad ${_nusim_build_branch_nomad} ${_nusim_build_checkout_nomad}
+            newStep 'gitCloneOrUpdate' 'refinery'
+            if [ ${useN} == "true" ]; then
+                newStep 'gitCloneOrUpdate' 'nomad'
+                cmdConcat "--nomad-path=${tmp}/repo/nomad"
             fi
-            gitCloneOrUpdate ${_git_GitHub_translations} ${_nusim_tmppath}/repo translations ${_nusim_build_branch_translations} ${_nusim_build_checkout_translations}
-            gitCloneOrUpdate ${_git_GitHub_refinery} ${_nusim_tmppath}/repo refinery ${_nusim_build_branch_refinery} ${_nusim_build_checkout_refinery}
+            if [ ${useT} == "true" ]; then
+                newStep 'gitCloneOrUpdate' 'translations'
+            else
+                cmdConcat "--no-latin"
+            fi
 
-            cd ${_nusim_sugar_mango}
+            cd ${mango}
             gitMango postCheckout
 
-            local cmd="nusim package:create:ps:upgrade -e dev"
-            cmd="${cmd} --mango-path ${_nusim_sugar_mango}"
-            cmd="${cmd} --baseline-path ${pathToBaseBuildPack}"
-            if [ "${_nusim_build_with_nomad}" == 'true' ]; then
-                cmd="${cmd} --nomad-path=${_nusim_tmppath}/repo/nomad"
-            fi
-            cmd="${cmd} --translations-path ${_nusim_tmppath}/repo/translations"
-            cmd="${cmd} --refinery-path ${_nusim_tmppath}/repo/refinery"
-            cmd="${cmd} --build-path ${_nusim_tmppath}/builds/${_nusim_build_upgrade_number}"
-            cmd="${cmd} --build-number ${_nusim_build_upgrade_number}"
-            cmd="${cmd} --sugar-version ${_nusim_sugar_version} --sugar-flavor ${_nusim_sugar_flavor}"
+            nusim -V
             if [ "$ME" == 'vagrant' ]; then
                 cmd="echo \"${cmd}\" | sudo su -"
             fi
-
-            nusim -V
-            tailPidCmd "${cmd}" '/tmp/nusim/nusim.log'
+            tailPidCmd "${cmd}" "${tmp}/nusim.log"
 
             local sleepCount=0
-            while [ -z "$(docker ps -a | grep build_core_install_)" -a $sleepCount -lt 300 ]; do
+            while [ -z "$(docker ps -a | grep build_core_)" -a $sleepCount -lt 300 ]; do
                 sleep $sleepUnit
                 let sleepCount=sleepCount+sleepUnit
             done
 
-            cmd="docker logs -f ${dockerImageName}"
-            if [ "$ME" == 'vagrant' ]; then
-                secho "echo '${cmd} &' | sudo su -" 'menu'
-                echo "${cmd}" | sudo su -
-            else
-                secho "${cmd}" 'menu'
-                ${cmd}
-            fi
+            newStep 'logDocker'
 
             wait $cmdPID
             sleep $sleepUnit
             secho "nusim command finished" 'menu'
 
-            nusimZipBuild ${_nusim_build_upgrade_number}
-            nusimCpFullBuild ${_nusim_build_upgrade_number}
+            nusimZipBuild ${upgradePack}
+            nusimCpFullBuild ${upgradePack}
 
             secho "/** Finished Create Upgrade Pack **/" 'menu'
         ;;
 
         'deployInstallPack')
-            local cmd="nusim package:deploy:ps:install -e dev"
-#            local installPack=$(ls ${_nusim_tmppath}/builds/refinery/${_nusim_build_number} | grep -oP "^[Sa-z]*.[\.|0-9]*.zip$")
+#            local installPack=$(ls ${tmp}/builds/refinery/${build} | grep -oP "^[Sa-z]*.[\.|0-9]*.zip$")
+            cmdConcat "package:deploy:${scope}:install -e dev"
+            cmdConcat "--package-zip ${tmp}/builds/zip/${build}.zip"
+            cmdConcat "--relative-path ${sugarName}"
+            cmdConcat "--license-key ${license}"
+            newStep 'dbKeyUser'
 
-            cmd="${cmd} --package-zip ${_nusim_tmppath}/builds/zip/${_nusim_build_number}.zip"
-            cmd="${cmd} --relative-path ${_nusim_build_number}${_nusim_sugar_name}"
-            cmd="${cmd} --db-user ${_nusim_sugar_db_user} --db-pass ${_nusim_sugar_db_pass}"
-
-            if [ ${_nusim_sugar_db_key} == 'oracle' ]; then
-                sqlPlusCreateUser ${_nusim_sugar_db_user} ${_nusim_sugar_db_pass}
-                cmd="${cmd} --db-type ${_db_oracle_type} --db-name ${_db_oracle_setRoot_host}/orcl"
-                cmd="${cmd} --db-host ${_db_oracle_connect_host} --db-port ${_db_oracle_port}"
-            else
-                cmd="${cmd} --db-type ${_db_mysql_type} --db-name turbinado_${_nusim_build_number}"
-                cmd="${cmd} --db-host ${_db_mysql_connect_host} --db-port ${_db_mysql_port}"
-            fi
-            cmd="${cmd} --license-key ${_nusim_sugar_license}"
-
+            nusim -V
             if [ "$ME" == 'vagrant' ]; then
                 cmd="echo \"${cmd}\" | sudo su -"
             fi
-
-            nusim -V
-            tailPidCmd "${cmd}" '/tmp/nusim/nusim.log'
+            tailPidCmd "${cmd}" "${tmp}/nusim.log"
             wait $cmdPID
             sleep $sleepUnit
             secho "nusim command finished" 'menu'
+            
+            newStep 'gitInit' ${scope}
         ;;
 
         'deployUpgradePack')
-            local cmd="nusim package:deploy:ps:upgrade -e dev --relative-path ${_nusim_build_number}${_nusim_sugar_name}"
-            cmd="${cmd} --package-zip ${_nusim_tmppath}/builds/zip/${_nusim_build_upgrade_to}.zip"
+            cmdConcat "package:deploy:${scope}:upgrade -e dev --relative-path ${sugarName}"
+            cmdConcat "--package-zip ${tmp}/builds/zip/${upgradePack}.zip"
+
+            if [ "$scope" == 'core' ]; then
+                cmdConcat "--silent-upgrader ${tmp}/builds/refinery/${build}/silentUpgrade-PRO-${configv}.zip"
+            fi
 
             if [ "$ME" == 'vagrant' ]; then
                 cmd="echo \"${cmd}\" | sudo su -"
             fi
 
             nusim -V
-            tailPidCmd "${cmd}" '/tmp/nusim/nusim.log'
+            tailPidCmd "${cmd}" "${tmp}/nusim.log"
             wait $cmdPID
             sleep $sleepUnit
+            secho "nusim command finished" 'menu'
         ;;
 
         'fullTest')
@@ -227,17 +220,119 @@ newsim()
     drawOptionDone
 }
 
+newStep() {
+    case ${1} in
+        'setVar')
+            cmd="nusim"
+            setYmlVal "_nusim_path_tmp" 'tmp'
+            setYmlVal "_nusim_path_mango" 'mango'
+            setYmlVal "_nusim_path_root" 'rootPath'
+        
+            setYmlVal "_nusim_build_db_key" 'dbKey'
+            setYmlVal "_db_${dbKey}_type" 'dbType'
+            setYmlVal "_db_${dbKey}_port" 'dbPort'
+            setYmlVal "_db_${dbKey}_connect_host" 'dbHost'
+            setYmlVal "_db_${dbKey}_connect_user" 'dbUser'
+            setYmlVal "_db_${dbKey}_connect_pass" 'dbPass'
+        
+            setYmlVal "_nusim_build_license" 'license'
+            setYmlVal "_nusim_build_suffix" 'suffix'
+            setYmlVal "_nusim_build_admin_user" 'adminName'
+            setYmlVal "_nusim_build_admin_pass" 'adminPass'
+        
+            setYmlVal "_nusim_build_recipe" 'build'
+            setYmlVal "_nusim_recipes_${build}_version" 'version'
+            setYmlVal "_nusim_recipes_${build}_configv" 'configv'
+            setYmlVal "_nusim_recipes_${build}_flavor" 'flavor'
+            setYmlVal "_nusim_recipes_${build}_upgrade" 'upgrade'
+
+            setYmlVal "_nusim_recipes_${build}_use_mangoCore" 'useMC'
+            setYmlVal "_nusim_recipes_${build}_use_nomad" 'useN'
+            setYmlVal "_nusim_recipes_${build}_use_translations" 'useT'
+            setYmlVal "_nusim_recipes_${build}_use_initGit" 'useG'
+            setYmlVal "_nusim_recipes_${build}_use_log_docker" 'useD'
+
+            sugarName="${build}${suffix}"
+            upgradePack="${build}_${upgrade}"
+            if [ "${useMC}" == "true" ]; then
+                mango="${tmp}/repo/mangoCore"
+            fi
+
+            if [ ${dbKey} == 'oracle' ]; then
+                setYmlVal "_db_${dbKey}_createuser_user" 'dbUser'
+                setYmlVal "_db_${dbKey}_createuser_pass" 'dbPass'
+            fi
+        ;;
+
+        'gitCloneOrUpdate')
+            setYmlVal "_git_GitHub_${2}" 'gitHubRepo'
+            setYmlVal "_nusim_recipes_${build}_${2}_branch" 'branchName'
+            setYmlVal "_nusim_recipes_${build}_${2}_commit" 'commit'
+            gitCloneOrUpdate ${gitHubRepo} "${tmp}/repo" ${2} ${branchName} ${commit}
+        ;;
+    
+        'dbKey')
+            cmdConcat "--db-type ${dbType} --db-host ${dbHost} --db-port ${dbPort}"
+            if [ ${dbKey} == 'oracle' ]; then
+                sqlPlusCreateUser ${dbUser} ${dbPass}
+                cmdConcat "--db-name ${_db_oracle_setRoot_host}/orcl"
+            else
+                cmdConcat "--db-name ${build,,}_${suffix}"
+            fi
+        ;;
+
+        'dbKeyUser')
+            newStep 'dbKey'
+            cmdConcat "--db-user ${dbUser} --db-pass ${dbPass}"
+        ;;
+
+        'logDocker')
+            local dockerImageName="build_core_${build}"
+            setYmlVal "_nusim_recipes_${build}_use_log_docker"
+            if [ ${ymlVal} == "true" ]; then
+                local cmd="docker logs -f ${dockerImageName}"
+                if [ "$ME" == 'vagrant' ]; then
+                    secho "echo '${cmd} &' | sudo su -" 'menu'
+                    echo "${cmd}" | sudo su -
+                else
+                    secho "${cmd}" 'menu'
+                    ${cmd}
+                fi
+            fi
+        ;;
+
+        'gitInit')
+            if [ ${useG} == "true" ]; then
+                cd "${_nusim_path_root}/${sugarName}"
+                rm -rf .git
+                pwd
+                gitConfig "initSugarBuild" "- ${2} install pack ${build}"
+                chown vagrant:vagrant -R .git
+            fi
+        ;;
+
+        *)
+            break
+            ;;
+    esac
+}
+
+cmdConcat() {
+    cmd="${cmd} ${1}"
+}
+
 nusimZipBuild() {
-    cd ${_nusim_tmppath}/builds/${1}
+    cd ${tmp}/builds/${1}
     zip -rq ${1}.zip ./
-    mkdir -p ${_nusim_tmppath}/builds/zip
-    mv ${1}.zip ${_nusim_tmppath}/builds/zip
-    ls ${_nusim_tmppath}/builds/zip/${1}.zip
+    mkdir -p ${tmp}/builds/zip
+    mv ${1}.zip ${tmp}/builds/zip
+    ls ${tmp}/builds/zip/${1}.zip
 }
 
 nusimCpFullBuild() {
-    mkdir -p ${_nusim_tmppath}/builds/refinery/${1}
-    cp ${_nusim_tmppath}/repo/refinery/build/* ${_nusim_tmppath}/builds/refinery/${1}
-    secho "ls -la ${_nusim_tmppath}/builds/refinery/${1}" 'menu'
-    ls -la ${_nusim_tmppath}/builds/refinery/${1}
+    mkdir -p ${tmp}/builds/refinery/${1}
+    cp ${tmp}/repo/refinery/build/* ${tmp}/builds/refinery/${1}
+    secho "ls -la ${tmp}/builds/refinery/${1}" 'menu'
+    ls -la ${tmp}/builds/refinery/${1}
+    rm -rf ${tmp}/repo/refinery/build/
 }